python/level1_single_api/9_amct/amct_onnx/requirements.txt
opencv-python
numpy
onnxruntime == 1.4.0
Pillow
